+++ examples/simple-js/index.js
@@ -1,5 +1,9 @@
 const { Numbers } = require("humanify-numbers");
 
-console.log("1234 ➜",Numbers.stringify(1234));
-console.log("1234 ➜",Numbers.stringify(1234, { format: 'byte' }), '(byte option)');
-console.log("1234 ➜",Numbers.stringify(1234, { precision: 3 }), '(precision 3 option)');
+console.log('1234 ➜', Numbers.stringify(1234)) // output: 1234 ➜ 1.2K
+
+console.log('1234 ➜', Numbers.stringify(1234, 
+                         { style: 'storage' })) // output: 1234 ➜ 1.2KB (style=storage)
+
+console.log('1234 ➜', Numbers.stringify(1234, 
+                           { precision: 3 })) // output: 1234 ➜ 1.234K (precision=3)