+++ examples/simple-ts/src/index.ts
@@ -3,7 +3,7 @@ import { Numbers } from 'humanify-numbers'
 console.log('1234 ➜', Numbers.stringify(1234)) // output: 1234 ➜ 1.2K
 
 console.log('1234 ➜', Numbers.stringify(1234, 
-                         { format: 'byte' })) // output: 1234 ➜ 1.2KB (byte)
+                         { style: 'storage' })) // output: 1234 ➜ 1.2KB (style=storage)
 
 console.log('1234 ➜', Numbers.stringify(1234, 
-                           { precision: 3 })) // output: 1234 ➜ 1.234K (precision 3)
+                           { precision: 3 })) // output: 1234 ➜ 1.234K (precision=3)