+++ lib/src/index.ts
@@ -4,18 +4,18 @@
 // Developed by simplyCoders, 2021.
 
 // ----------------------------------------------
-// Supported formats.
-const formatsAvailable = ['number', 'byte', 'percent']
-interface IFormats {
-    [key: string]: any
+// Supported styless.
+const stylesAvailable = ['number', 'storage', 'percent', 'byte']
+interface IStyles {
+  [key: string]: any
 }
-const formats: IFormats = {
+const styles: IStyles = {
   number: {
     suffix: ['', 'K', 'M', 'B', 'T'],
     min: 10 ** (-3),
     max: 10 ** 15,
   },
-  byte: {
+  storage: {
     suffix: ['B', 'KB', 'MB', 'GB', 'TB', 'PB'],
     min: 10 ** (0),
     max: 10 ** 18,
@@ -31,7 +31,7 @@ const formats: IFormats = {
 // Supported math functions.
 const mathsAvailable = ['floor', 'round', 'ceil']
 interface IMaths {
-    [key: string]: any
+  [key: string]: any
 }
 const maths: IMaths = {
   floor: Math.floor,
@@ -44,7 +44,7 @@ const maths: IMaths = {
 export class Numbers {
   static stringify(value: number, options?: any) {
     // defaults
-    const defaultOpts = { format: 'number', precision: 1, math: 'floor' }
+    const defaultOpts = { style: 'number', precision: 1, math: 'floor' }
 
     // validate value
     if (typeof value !== 'number') {
@@ -54,9 +54,19 @@ export class Numbers {
     // analyze options
     const opts = defaultOpts
     if ((typeof options !== 'undefined')
+      && (typeof options.style === 'string')
+      && (stylesAvailable.includes(options.style))) {
+      opts.style = options.style
+    }
+
+    if ((typeof options !== 'undefined') // for backward compatability
       && (typeof options.format === 'string')
-      && (formatsAvailable.includes(options.format))) {
-      opts.format = options.format
+      && (stylesAvailable.includes(options.format))) {
+      opts.style = options.format
+    }
+
+    if (opts.style === 'byte') { // for backward compatability
+      opts.style = 'storage'
     }
 
     if ((typeof options !== 'undefined')
@@ -72,28 +82,28 @@ export class Numbers {
       opts.math = options.math
     }
 
-    const val = (opts.format === 'percent') ? (value * 100) : value
+    const val = (opts.style === 'percent') ? (value * 100) : value
 
     // end cases
     if (val === 0) {
-      const zSuffix = formats[opts.format].suffix[0]
+      const zSuffix = styles[opts.style].suffix[0]
       const zeroVal = (0).toString()
       return zeroVal + zSuffix
     }
 
     const absVal = Math.abs(val)
-    if ((absVal < formats[opts.format].min) || (absVal > formats[opts.format].max)) {
+    if ((absVal < styles[opts.style].min) || (absVal > styles[opts.style].max)) {
       return 'NaN'
     }
 
-    if (opts.format === 'byte' && (!Number.isInteger(val) || val < 0)) {
+    if (opts.style === 'storage' && (!Number.isInteger(val) || val < 0)) {
       return 'NaN'
     }
 
     // process the value
     let numberIndex = Math.floor(Math.log(absVal + 0.00001) / Math.log(10) / 3)
     numberIndex = (numberIndex < 0) ? 0 : numberIndex
-    const suffix = formats[opts.format].suffix[numberIndex]
+    const suffix = styles[opts.style].suffix[numberIndex]
 
     const numDigitsToRemove = numberIndex * (-3) + opts.precision
     const digitsToKeep = val * (10 ** numDigitsToRemove)